test: add DB state assertions for transaction sync tests
