+++ pkg/storage/internal/testabilities/assertions_sync_chunk.go
@@ -1,21 +1,13 @@
 package testabilities
 
 import (
-	"context"
-	"maps"
 	"testing"
 
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/storage/entity"
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/wdk"
 	testvectors "github.com/bsv-blockchain/universal-test-vectors/pkg/testabilities"
-	"github.com/go-softwarelab/common/pkg/seq"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-type StorageReader interface {
-	FindKnownTx(ctx context.Context, txID string) (*entity.KnownTx, error)
-}
 
 type SyncAssertion interface {
 	Chunk(chunk *wdk.SyncChunk) SyncChunkAssertion
@@ -62,17 +54,6 @@ type ProvenTxAssertion interface {
 	HasMerklePath() ProvenTxAssertion
 }
 
-type DBStateAssertion interface {
-	HasKnownTXs(txIDs ...string) DBStateAssertion
-	HasKnownTX(txID string) KnownTxAssertion
-}
-
-type KnownTxAssertion interface {
-	WithStatus(state wdk.ProvenTxReqStatus) KnownTxAssertion
-	IsMined() KnownTxAssertion
-	HasRawTx() KnownTxAssertion
-}
-
 type syncAssertion struct {
 	testing.TB
 }
@@ -96,10 +77,7 @@ func (s *syncAssertion) DBState(storage StorageReader) DBStateAssertion {
 	s.Helper()
 	require.NotNil(s, storage, "Expected storage to be not nil")
 
-	return &dbStateAssertion{
-		TB:      s.TB,
-		storage: storage,
-	}
+	return ThenDBState(s, storage)
 }
 
 func (s *syncChunkAssertion) WithoutError(err error) ValidSyncChunkAssertion {
@@ -278,76 +256,4 @@ func (p *proveTxAssertion) HasMerklePath() ProvenTxAssertion {
 	p.parent.Helper()
 	assert.NotEmpty(p.parent, p.tx.MerklePath, "Expected tx to have a non-empty MerklePath")
 	return p
-}
-
-type dbStateAssertion struct {
-	testing.TB
-	storage StorageReader
-}
-
-func (d *dbStateAssertion) HasKnownTXs(txIDs ...string) DBStateAssertion {
-	d.Helper()
-
-	missingTXs := map[string]struct{}{}
-
-	for _, txID := range txIDs {
-		knownTx, err := d.storage.FindKnownTx(d.Context(), txID)
-		require.NoError(d.TB, err, txID)
-
-		if knownTx == nil {
-			missingTXs[txID] = struct{}{}
-		}
-	}
-
-	if len(missingTXs) != 0 {
-		missingIDs := seq.Collect(maps.Keys(missingTXs))
-		assert.Failf(d, "Expected to find all the transactions", "missing transaction IDs: %v", missingIDs)
-	}
-
-	return d
-}
-
-func (d *dbStateAssertion) HasKnownTX(txID string) KnownTxAssertion {
-	d.Helper()
-
-	knownTx, err := d.storage.FindKnownTx(d.Context(), txID)
-	require.NoError(d.TB, err, txID)
-
-	if knownTx == nil {
-		require.Failf(d, "Expected to find the transaction", "transaction ID: %s", txID)
-		return nil
-	}
-
-	assert.Equal(d, txID, knownTx.TxID, "Expected known transaction to have the same TxID as the one requested")
-
-	return &knownTxAssertion{
-		TB:      d.TB,
-		knownTx: knownTx,
-	}
-}
-
-type knownTxAssertion struct {
-	testing.TB
-	knownTx *entity.KnownTx
-}
-
-func (d *knownTxAssertion) WithStatus(state wdk.ProvenTxReqStatus) KnownTxAssertion {
-	d.Helper()
-	assert.Equal(d, state, d.knownTx.Status, "Expected known transaction to have the status %s", state)
-	return d
-}
-
-func (d *knownTxAssertion) IsMined() KnownTxAssertion {
-	d.Helper()
-	assert.NotNil(d, d.knownTx.BlockHeight)
-	assert.NotEmpty(d, d.knownTx.MerklePath)
-	assert.NotEmpty(d, d.knownTx.MerkleRoot)
-	assert.NotEmpty(d, d.knownTx.BlockHash)
-	return d
-}
-
-func (d *knownTxAssertion) HasRawTx() KnownTxAssertion {
-	d.Helper()
-	assert.NotEmpty(d, d.knownTx.RawTx, "Expected known transaction to have a non-empty RawTx")
-	return d
 }
